+++ MetroSim/classes/Udalost.cs
@@ -2,11 +2,11 @@
 {
     class Udalost
     {
-        public int kdy;
+        public float kdy;
         public Proces kdo;
         public TypUdalosti co;
 
-        public Udalost(int kdy, Proces kdo, TypUdalosti co)
+        public Udalost(float kdy, Proces kdo, TypUdalosti co)
         {
             this.kdy = kdy;
             this.kdo = kdo;
